pyproject.toml
[project]
name = "intelligent-message-routing"
version = "0.1.0"
description = "Triage and route customer messages with PII redaction and policy-grounded draft responses"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
    "pytest>=7.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
